Wait 5 milliseconds before doing Simple Pairing auto-accept
The BITE tester is again really stupid and it can't handle dongles
that can do the DH key exchange quickly. For example the Broadcom
dongles use a fast processors and an optimized DH algorithm. So in
the end they are way too fast.

To make this work, use a 5 milliseconds delay before replying to
the auto-accept case of Simple Pairing.