Fix potential AVDTP reference leak
If the AVDTP connection has just recently disconnected but not yet been
notified to the mainloop then we need to clean the unix.c avdtp session
reference within start_discovery() instead of doing it when the unix
client disconnects. This is because when the AVDTP disconnection finally
gets reported in the next mainloop iteration the unix.c reference wouldn't
get cleared as is intended (since unix.c doesn't have any callbacks at
this point).