Fix AVDTP ABORTING state transition
The transition to ABORTING state when acting as initiator should happen
when we receive the ABORT reply and not when we send the initial ABORT
command. The reply handling is already correct but since there's a state
change also in the sending part the later state change request would fail.
This patch fixes the issue by removing the state change when sending the
ABORT command.

Change-Id: I9cf9f27924b7d348523125ddccff7c8ecebacf93